+++ Coraza.h
@@ -12,7 +12,7 @@ public:
   void setArmadura(int);
   void setLanzas(int);
 
-  double defensa();
-  double ataque();
+  int defensa();
+  int ataque();
 
 };
